% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fukan_Keywords_to_list}
\alias{fukan_Keywords_to_list}
\title{Transform fukan system keyword files to the input object for heatmaps}
\usage{
fukan_Keywords_to_list(path, report = FALSE)
}
\arguments{
\item{path}{a directory having the "mission.keyword.tsv" files}

\item{report}{whether to save the unified keyword report as .csv}
}
\description{
Transform fukan system keyword files to the input object for heatmaps
}
\examples{
fukan_keywords_to_list(path)
}
\keyword{clusters}
\keyword{read,}
