% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{keywords_to_list}
\alias{keywords_to_list}
\title{Transform topic model reports to the input object for heatmaps}
\usage{
keywords_to_list(a_data_frame)
}
\arguments{
\item{a_data_frame}{a dataframe created with topic_model_5.R}
}
\description{
Transform topic model reports to the input object for heatmaps
}
\examples{
keyword_to_list(a_data_frame)
}
\keyword{models}
\keyword{read,}
\keyword{topic}
