% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{unigrams}
\alias{unigrams}
\title{Transform list of n_grams to unigrams}
\usage{
unigrams(x)
}
\arguments{
\item{x}{a dataframe of keywords}
}
\description{
Transform list of n_grams to unigrams
}
\examples{
keyword_to_list(a_data_frame)
}
\keyword{models}
\keyword{read,}
\keyword{topic}
